light refactoring and commenting
