+++ .vscode/settings.json
@@ -15,6 +15,8 @@
     "htmlescape",
     "indexmap",
     "itertools",
+    "MDSL",
+    "micropip",
     "msdf",
     "nonblank",
     "Northbridge",
